Added authentication in journal

+++ src/main/java/com/learner/journalApp/service/JournalEntryService.java
@@ -1,11 +1,6 @@
 package com.learner.journalApp.service;
 
 
-// BEST PRACTICE-
-// Controller calls -> service
-// Service calls -> repository
-// Controller ---> Service ---> Repository
-
 import com.learner.journalApp.entity.JournalEntry;
 import com.learner.journalApp.entity.User;
 import com.learner.journalApp.repository.JournalEntryRepository;
@@ -37,11 +32,10 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            user.setUserName(null);         // error to learn Transactional
             userService.saveEntry(user);
         }catch (Exception e){
             log.error("Exception", e);
-            throw new RuntimeException("An exception accures while Saving Entry: ", e);
+            throw new RuntimeException("An exception occurs while Saving Entry: ", e);
         }
     }
 
@@ -57,11 +51,20 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));   // it will check all journal entries inside list and remove if id is present
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));   // it will check all journal entries inside list and remove if id is present
+            if(removed){
+                userService.saveEntry(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("At error occur while deleting the journal");
+        }
+        return removed;
     }
 
 }
